--- docs/work_logs/2026-01-11_work_log.md
@@ -1,43 +1,0 @@
-# 2026-01-11 작업 로그 및 계획 (Work Log)
-
-## ✅ 1/10 토 작업한 내용 (Accomplished)
-
-
-### 0. REAd ME 정리
-
-### 1. 채팅창 기반 테스트 시스템 복구 (`/t` 커맨드)
-- **기능 구현**: 터미널 스크립트 실행 없이 채팅창에서 바로 테스트 수행 가능.
-- **커맨드 형식**: `/t [목적],[시나리오],[기대값] 실제질문`
-- **안정성 강화**:
-    - `Supervisor` 노드에서 정규식(`re`)을 사용해 `[ ]` 포맷을 유연하게 파싱하도록 개선.
-    - 파싱 실패 시 서버 크래시 방지(Exception Handling).
-    - `import time` 누락 등의 버그 수정 완료.
-
-### 2. 자동 리포트 생성 시스템 구축
-- **파일 위치**: `backend/test_reports/test_YYYY-MM-DD.md`
-- **데이터 적재**: 11개 컬럼(시간, 목적, 시나리오, 소요시간, 비용 등) 데이터를 자동으로 마크다운 표 형태로 저장.
-- **비용 산출**: Input($1.75)/Output($14.00) 토큰 단가 기준 자동 계산 로직 적용.
-
-### 3. Writer 노드 프롬프트 두줄 수정
-- **앵무새 방지**: 검색 결과가 없어도("Small Talk" 상황) DB 타령을 하지 않고 자연스럽게 대화하도록 규칙 수정.
-- **주석 관리**: 레거시(원본) 규칙을 파이썬 주석으로 보존하고 새로운 유연한 규칙 적용.
-
----
-
-## 📅 1/11 일 할 일 (Next Steps: Prompt Engineering & Redesign)
-
-### 1. Supervisor & Interviewer 두뇌 교체 (로직 재설계)
-- **목표**: "본격 프롬프팅."
-- **Supervisor 개선**:
-    - **Current**: 사용자 발화 1개만 보고 판단 (문맥 파악 불가).
-    - **Plan**: 이전 대화 문맥(`context`)을 함께 참조하여 라우팅 결정 (Context-Aware).
-- **Interviewer 개선**:
-    - **Current**: 텍스트 줄글 요약 (비구조적).
-    - **Plan**: **Slot Filling (슬롯 채우기)** 방식 도입. (`Image`, `Season`, `Brand` 등 구조화된 딕셔너리로 관리).
-
-### 2. 프롬프트 엔지니어링 (Persona)
-- **질문 고도화**: 사용자가 모호하게 말했을 때("차가운 거"), 구체적인 이미지를 잡아내는 유도 질문 로직 강화.
-
-### 3. 정리
-- **Git Upload**: 지금까지 작업본 커밋 및 푸시.
-- **휴식**: (스몰슬립 예정)
